pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "paper-parser"
version = "0.1.0"
description = "A CLI tool for searching, downloading, parsing, and summarizing academic papers."
readme = "README.md"
authors = [
    { name = "kaihang", email = "your.email@example.com" }
]
dependencies = [
    "requests",
    "click",
    "PyYAML",
    "arxiv",
    "rapidfuzz",
]
requires-python = ">=3.8"

[project.scripts]
paper-parser = "paper_parser.cli:cli"
pp = "paper_parser.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["paper_parser*"]
